refactor: move Lotto.kt to models directory and move few validation to Lotto class from Inputvalidation
