--- src/main/kotlin/lotto/Lotto.kt
@@ -1,11 +1,0 @@
-package lotto
-
-class Lotto(private val numbers: List<Int>) {
-    init {
-        require(numbers.size == 6) { "[ERROR] Lotto must contain exactly 6 numbers." }
-    }
-
-    override fun toString(): String {
-        return numbers.toString()
-    }
-}
